CLI Routing - Dumbed down architecture
Architecture dumbed down with the guidance of Twitter's "Effective Scala" guide.
Functions no longer return futures, leaving the burden of concurrency to the caller.
The progress tracking framework has been reworked to be more flexible and decentralized. It now only defines what a generic ProgressHook is, and leaves it up to the hook targets and hook providers to coordinate type inference and binding.
Fixed many compilation issues, primarily Type inference.