Download Mechanism - Some Foundation
Had to redefine many of the hook points, which has broken the InstallModpack command for now. That will be fixed in this branch before merging.
HTTP code written, and the installation routine has been restructured a little to make progress tracking more granular.
Still need to calla the HTTP code from the routine, and to fix the hooks in the command.