Download Mechanism - Dependency and SDK Changes
Removed Spring from dependencies and the code which depended on it.
Reverted to using Java 17 instead of 8, due to the useful features available in 17.
End-Users will now need to download Java SE 17 from Oracle's site, which is different from the default Java download site that the average person would find through a search engine.