Sprint 1 - Complete
The time has come to end Sprint 1. Due to the drastic reduction in my availability to work on this project, one of the major features, Profile Injection, has been postponed to a future sprint. The Download Mechanism works and has been tested, as well as the currently implemented PicoCLI command. It currently uses default input data, which will be replaced with actual web service requests in the next client CLI sprint.